refactor(cli): collapse twin planners and finish Actions helpers
Address the thermo-nuclear review on the nesting PR: share axis planning
and registry selection, route every bench-suite cell exit through one
summary path, keep suite-task discovery free of HTML presentation, lift
logInfo into actions-log, and split workflow drift gates by concern.
